Drop GPS and location usage
GPS is being jammed; coordinates are inaccurate.
As a result, we will be relying on a different solution.

Initially, we wanted to use GPS to render a Radar relative to our position.
Now, we'll be using BLE RSSI and Sound Signaling.

If we want to use Location again, we can revert this change.